+++ crypto/entropy.go
@@ -3,7 +3,7 @@ package crypto
 import "bytes"
 
 // Increase entropy by multiple hashing.
-func RaiseEntropy(info, salt []byte, bits uint) []byte {
+func RaiseEntropy(info, salt []byte, bits uint64) []byte {
 	lim := uint64(1 << bits)
 	for i := uint64(0); i < lim; i++ {
 		info = NewSHA256(bytes.Join(
